Keep child navigation within bundle budget
